requirements.txt
asyncpg==0.30.0
base_processor==0.0.2
colorama==0.4.6
datasketch==1.6.5
fastapi==0.115.12
langchain==0.3.23
langchain_chroma==0.2.2
langchain_community==0.3.21
langchain_core==0.3.51
langchain_ollama==0.3.1
langchain_openai==0.3.12
langgraph==0.3.27
monitoring==0.9.12
nltk==3.9.1
numpy==2.2.4
pandas==2.2.3
pydantic==2.11.3
python-dotenv==1.1.0
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.6.1
sentence_transformers==4.0.2
SQLAlchemy==2.0.40
streamlit==1.44.1
torch==2.6.0
tqdm==4.67.1
PyMuPDF==1.27.11
unstructured>=0.17.2